+++ packages/algob/src/lib/account.ts
@@ -3,9 +3,9 @@ import { Account as AccountSDK, Kmd, mnemonicToSecretKey, multisigAddress, Multi
 import * as fs from "fs";
 import YAML from "yaml";
 
+import { BuilderError } from "../errors/errors";
+import { ERRORS } from "../errors/errors-list";
 import CfgErrors, { ErrorPutter } from "../internal/core/config/config-errors";
-import { BuilderError } from "../internal/core/errors";
-import { ERRORS } from "../internal/core/errors-list";
 import type { Account, AccountDef, HDAccount, KmdCfg, KmdWallet, MnemonicAccount, StrMap } from "../types";
 
 /**