+++ packages/runtime/test/src/interpreter/opcode-list.ts
@@ -217,11 +217,14 @@ describe("Teal Opcodes", function () {
 
   describe("Arg[N]", function () {
     const stack = new Stack<StackElem>();
-    const runtime = new Runtime([]);
-    const interpreter = new Interpreter();
+    let interpreter: Interpreter;
     const args = ["Arg0", "Arg1", "Arg2", "Arg3"].map(stringToBytes);
-    interpreter.runtime = runtime;
-    interpreter.runtime.ctx.args = args;
+
+    this.beforeAll(() => {
+      interpreter = new Interpreter();
+      interpreter.runtime = new Runtime([]);
+      interpreter.runtime.ctx.args = args;
+    });
 
     it("should push arg_0 from argument array to stack", function () {
       const op = new Arg(["0"], 1, interpreter);
@@ -2278,18 +2281,17 @@ describe("Teal Opcodes", function () {
   describe("Global Opcode", function () {
     const stack = new Stack<StackElem>();
     let interpreter: Interpreter;
+    this.beforeAll(() => {
+      // setup 1st account (to be used as sender)
+      const acc1: AccountStoreI = new AccountStore(123, { addr: elonAddr, sk: new Uint8Array(0) }); // setup test account
+      setDummyAccInfo(acc1);
 
-    // setup 1st account (to be used as sender)
-    const acc1: AccountStoreI = new AccountStore(123, { addr: elonAddr, sk: new Uint8Array(0) }); // setup test account
-    setDummyAccInfo(acc1);
-
-    before(() => {
       interpreter = new Interpreter();
       interpreter.runtime = new Runtime([acc1]);
       interpreter.runtime.ctx.tx = TXN_OBJ;
       interpreter.runtime.ctx.gtxs = [TXN_OBJ];
       interpreter.runtime.ctx.tx.apid = 1828;
-      interpreter.tealVersion = MaxTEALVersion; // set tealversion to latest (to support all global fields)
+      interpreter.tealVersion = 2; // set tealversion to latest (to support all global fields)
     });
 
     it("should push MinTxnFee to stack", function () {
@@ -2407,7 +2409,6 @@ describe("Teal Opcodes", function () {
 
   describe("StateFul Opcodes", function () {
     const stack = new Stack<StackElem>();
-    const interpreter = new Interpreter();
     const lineNumber = 0;
 
     // setup 1st account (to be used as sender)
@@ -2418,14 +2419,17 @@ describe("Teal Opcodes", function () {
     const acc2 = new AccountStore(123, { addr: johnAddr, sk: new Uint8Array(0) });
     setDummyAccInfo(acc2);
 
-    const runtime = new Runtime([acc1, acc2]);
-    interpreter.runtime = runtime; // setup runtime
+    let interpreter: Interpreter;
+    this.beforeAll(() => {
+      interpreter = new Interpreter();
+      interpreter.runtime = new Runtime([acc1, acc2]);
 
-    // setting txn object and sender's addr
-    interpreter.runtime.ctx.tx = {
-      ...TXN_OBJ,
-      snd: Buffer.from(decodeAddress(elonAddr).publicKey)
-    };
+      // setting txn object and sender's addr
+      interpreter.runtime.ctx.tx = {
+        ...TXN_OBJ,
+        snd: Buffer.from(decodeAddress(elonAddr).publicKey)
+      };
+    });
 
     describe("AppOptedIn", function () {
       it("should push 1 to stack if app is opted in", function () {